// src/types.ts
export interface ColourScheme {
    /**
     * This is the main color that represents the brand
     * or the most dominant color in the website's palette. 
     * It's used for major call-to-action (CTA) buttons, 
     * links, and headlines to draw attention.
     */
    primary: string,

    /**
     * Supports the primary color and is used for secondary buttons, 
     * highlights, and information boxes. 
     * The secondary color should complement the primary color.
     */
    secondary: string,

    /**
     * Lighter shade of the primary color
     * These are additional colors used sparingly to accentuate
     * or highlight parts of the website.
     * They can be used for icons,
     * info graphics, or to highlight key information.
     */
    accent: string,

    /**
     * A fixed light grey
     * These are usually neutral colors that
     * don't compete with the primary and secondary colors.
     */
    background: string,


    /**
     * A fixed dark grey for text
     * Besides the obvious black or white,
     * you might have specific colors for headings,
     * subheadings, and body text that fit within
     * the overall color scheme.
     */
    text: string,

    /**
     * Lighter shade of the secondary color
     * These include shades of gray, beige,
     * or off-white used for text, background,
     * borders, and shadows.
     * They help create hierarchy and contrast in the design.
     */
    neutral: string,

    palette: string[]
}

export enum ColourPalette {
    ACCENT = '--color-accent',
    BACKGROUND = '--color-background',
    NEUTRAL = '--color-neutral',
    PRIMARY = '--color-primary',
    SECONDARY = '--color-secondary',
    TEXT = '--color-text'
}


/**
 * Font
 */
export interface FamilyFont {
    name: string
    family: string
    css: string
    description?: string
}

export interface FontScheme {
    heading: FamilyFont
    body: FamilyFont
}

export enum FontFamily {
    HEADING = "--font-family-heading",
    BODY = "--font-family-body"
}
// ...
export enum ProjectMessageType {
    SET_TITLE = "SET_TITLE",
    SET_PAGE = "SET_PAGE",
    READ_PAGE = "READ_PAGE",
    DELETE_PAGE = "DELETE_PAGE",

    SET_COLOUR_SCHEME = "SET_COLOUR_SCHEME",
    SET_FONT_SCHEME = "SET_FONT_SCHEME",
}